requirements.txt
ruff
mypy
pytest
BeautifulSoup4
requests
polars
loguru
pydantic
pillow